Restructuring code for connections and channel models
