+++ qunetsim/objects/__init__.py
@@ -1,10 +1,7 @@
-from .classical_storage import ClassicalStorage
+from qunetsim.objects.storage.classical_storage import ClassicalStorage
 from .logger import Logger
 from .message import Message
-from .packet import Packet
-from .quantum_storage import QuantumStorage
+from qunetsim.objects.packets.packet import Packet
 from .qubit import Qubit
-from .quantum_connection import QuantumConnection, Fibre, BinaryErasure
-from .classical_connection import ClassicalConnection
-from .routing_packet import RoutingPacket
+from qunetsim.objects.connections.quantum_connection import QuantumConnection, Fibre, BinaryErasure
 from .daemon_thread import DaemonThread
