+++ qunetsim/objects/connections/channel_models/fibre.py
@@ -1,47 +1,7 @@
-import random
-
-
-class QuantumConnection(object):
-    """
-    An object that stores quantum connection details
-    """
-
-    def __init__(self, receiver_id, model=None):
-        self._receiver_id = receiver_id
-        if model is None:
-            self._model = Fibre()  # Defaults to fibre model
-        else:
-            self._model = model
-
-    @property
-    def receiver_id(self):
-        """
-        Receiver ID
-        """
-        return self._receiver_id
-
-    @property
-    def model(self):
-        """
-        Channel model
-
-        Returns:
-            (object) : An object containing model characteristics
-        """
-        return self._model
-
-    @model.setter
-    def model(self, model):
-        """
-        Set the channel model
-
-        Args
-            model (object) : An object containing model characteristics and parameters
-        """
-        self._model = model
-
-
 class Fibre(object):
+    """
+    TODO: Add docstring
+    """
 
     def __init__(self, length=0.0, alpha=0.0):
 
@@ -129,58 +89,6 @@ class Fibre(object):
             (object) : Modified qubit
         """
         if random.random() > self.transmission_p:
-            if qubit is not None:
-                qubit.release()
-            return None
-        else:
-            return qubit
-
-
-class BinaryErasure(object):
-
-    def __init__(self, probability=0.0):
-        if not isinstance(probability, int) and not isinstance(probability, float):
-            raise ValueError("Erasure probability must be float or int")
-        elif probability < 0 or probability > 1:
-            raise ValueError("Erasure probability must lie in the interval [0, 1]")
-        else:
-            self._P = probability
-
-    @property
-    def erasure_probability(self):
-        """
-        Probability of erasure of qubit
-
-        Returns
-            (float) : Probability that a qubit is erased during transmission
-        """
-        return self._P
-
-    @erasure_probability.setter
-    def erasure_probability(self, probability):
-        """
-        Set the erasure probability of the channel
-
-        Args
-            probability (float) : Probability that a qubit is erased during transmission
-        """
-        if not isinstance(probability, int) and not isinstance(probability, float):
-            raise ValueError("Erasure probability must be float or int")
-        elif probability < 0 or probability > 1:
-            raise ValueError("Erasure probability must lie in the interval [0, 1]")
-        else:
-            self._P = probability
-
-    def qubit_func(self, qubit):
-        """
-        Function to modify the qubit based on channel properties
-        In this case - Returns None if qubit is erased, otherwise returns the original qubit
-        Required in all channel models
-
-        Returns
-            (object) : Modified qubit
-        """
-        if random.random() > (1.0 - self.erasure_probability):
             if qubit is not None:
                 qubit.release()
             return None
