+++ qunetsim/objects/connections/connection.py
@@ -1,0 +1,7 @@
+"""
+TODO: Complete this object
+"""
+
+class Connection(object):
+    def __init__(self, receiver_id):
+        self._receiver_id = receiver_id
